chore: updated comments from WK6 from Tetiana, moved TodoList and TodoListItem components to the TodoList directory, and updated prop references in TodoForm component.
