+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import './App.css';
-import TodoList from './features/TodoList';
+import TodoList from './features/TodoList/TodoList';
 import TodoForm from './features/TodoForm';
 
 function App() {
